move frontend types back into cost-insights
