+++ plugins/cost-insights/src/types/Currency.ts
@@ -14,9 +14,13 @@
  * limitations under the License.
  */
 
-export enum CurrencyType {
-  USD = 'USD',
-  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
-  Beers = 'BEERS',
-  IceCream = 'PINTS_OF_ICE_CREAM',
+/**
+ * @public
+ */
+export interface Currency {
+  kind: string | null;
+  label: string;
+  unit: string;
+  prefix?: string;
+  rate?: number;
 }
