+++ plugins/cost-insights/src/types/CurrencyType.ts
@@ -14,13 +14,9 @@
  * limitations under the License.
  */
 
-/**
- * @public
- */
-export interface Currency {
-  kind: string | null;
-  label: string;
-  unit: string;
-  prefix?: string;
-  rate?: number;
+export enum CurrencyType {
+  USD = 'USD',
+  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
+  Beers = 'BEERS',
+  IceCream = 'PINTS_OF_ICE_CREAM',
 }
